+++ packages/style/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fat-fuzzy/style
 
+## 0.1.43
+
+### Patch Changes
+
+- 67c7f21: Improve base schemas, config AJV options, a11y fixes
+
 ## 0.1.42
 
 ### Patch Changes