+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fat-fuzzy/ui
 
+## 0.1.99
+
+### Patch Changes
+
+- 67c7f21: Improve base schemas, config AJV options, a11y fixes
+
 ## 0.1.98
 
 ### Patch Changes